requirements.txt
python-dotenv==1.0.1
discord.py
discord
aiohttp==3.9.5
discord.py==2.4.0
table2ascii==1.1.3